평점 평균 계산 및 평점 모델 추가( 소수점 계산 처리)

+++ API/restfulAPI/dataTest/views.py
@@ -176,7 +176,7 @@ def challenge_content(request, pk):
         total_rate = rates.aggregate(total=Sum('rate'))['total']
         rate_count = rates.count()
 
-        average_rate = total_rate / rate_count
+        average_rate = int(total_rate / rate_count)
 
         update_data = {
             'rate': average_rate
